+++ src/utils/form/index.js
@@ -1,2 +1,3 @@
 export * from './formDataUtils';
 export * from './registrationFormUtils';
+export * from './formUtils';
