+++ src/utils/validations/index.js
@@ -1,2 +1,1 @@
-export * from './formUtils';
-export * from './formValidations';
+export * from './validations';
