--- src/utils/validations/validation.js
@@ -1,19 +1,0 @@
-import {
-  nameValidator,
-  ageValidator,
-  mobileValidator,
-  optionalEmailValidator,
-  addressValidator,
-  optionalMobileValidator,
-  bhopalAgeValidator,
-} from './formValidations';
-
-export default {
-  nameValidator,
-  ageValidator,
-  mobileValidator,
-  optionalEmailValidator,
-  addressValidator,
-  optionalMobileValidator,
-  bhopalAgeValidator,
-};
